fix(payments): use exact Decimal/NUMERIC money; fix wallet race; idempotent capped refunds
Replace float money with NUMERIC(12,2) + Python Decimal end to end (restaurant
prices/fees, order financials, wallet balances/transactions). Eliminates
rounding drift and the order(dollars)->payment(cents) unit mismatch. API
contract is unchanged (responses still serialize as JSON numbers).

- restaurant-svc: menu price, delivery_fee, minimum_order_amount -> NUMERIC(12,2);
  request schemas parse to exact Decimal; OpenSearch doc casts to float.
- order-svc: order financials -> NUMERIC(12,2); order-total math via a Decimal
  _money() helper; float() only at the JSONB snapshot and SNS event boundaries.
- payment-svc: wallet balance/transactions -> NUMERIC(12,2); credit/debit coerce
  to Decimal and lock the wallet row (SELECT ... FOR UPDATE) to close the
  read-check-write overdraft race; fixed cents<->dollars conversion.
- payment-svc refunds: add cumulative refunded_amount; cap to remaining
  refundable amount; optional idempotency_key replays the original refund
  (also forwarded to Stripe) instead of double-charging.
- Three Alembic 0002 migrations (non-destructive USING ::numeric(12,2)).

Verified against in-memory SQLite: exact arithmetic, overdraft rejection,
and the order total calculation.

+++ services/restaurant-svc/app/models/menu.py
@@ -1,7 +1,8 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import Boolean, DateTime, Float, ForeignKey, Integer, String, Text, func
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.dialects.postgresql import ARRAY, JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -40,7 +41,7 @@ class MenuItem(Base):
 
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    price: Mapped[float] = mapped_column(Float, nullable=False)
+    price: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     image_url: Mapped[str | None] = mapped_column(Text, nullable=True)
     category: Mapped[str] = mapped_column(String(100), nullable=False, default="")
